pyproject.toml

[tool.poetry]
name = "voicelab"
version = "2.0.0"
description = "Automated Reproducible Voice Analysis"
authors = ["David Feinberg <feinberg@mcmaster.ca>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9, <=3.11.3"
numpy = ">=1.22"
PyQt5 = ">=5.15.3"
QDarkStyle = "2.8.1"
praat-parselmouth = ">=0.4.1"
librosa = "0.9.2"
openpyxl = "3.0.6"
seaborn = "0.11.1"

[tool.poetry.group.dev.dependencies]
PyInstaller = "^5.10.0"
Sphinx = "^6.1.3"
sphinx-autodoc2 = "^0.4.2"
sphinx-material = "0.0.35"
pytest = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
voicelab = "voicelab.src.Voicelab.voicelab_script:main"
